% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_biomarkers_user.R
\name{compute_lipoprotein_fatty_acids}
\alias{compute_lipoprotein_fatty_acids}
\title{Compute fatty acids in each of the lipoprotein classes and subclasses}
\usage{
compute_lipoprotein_fatty_acids(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.
May either be raw field data output by the Table Exporter tool on the UK
Biobank Research Analysis Platform or data with column names corresponding
to biomarkers listed in \code{\link{nmr_info}}.}
}
\value{
a \code{data.frame} with the additional computed biomarkers.
}
\description{
{
Computes fatty acids in each lipoprotein subclass following the formula
derived by Belkadi \emph{et al.} 2026:

\deqn{\text{Fatty Acids} = \text{Cholesteryl Esters} + 2 \times Phospholipids + 3 \times Triglycerides}

Fatty acids in each lipoprotein class are computed as the sum of the fatty
acids in the consituent lipoprotein subclasses.
}
}
\details{
If your UK Biobank project only has access to a subset of biomarkers, then
this function will only return the subset of lipoprotein fatty acids that
can be computed from the biomarker data provided.

All biomarkers in the input data are also returned alongside the new
biomarkers computed by this function
}
\examples{
ukb_data <- ukbnmr::test_data # Toy example dataset for testing package
nmr <- compute_lipoprotein_fatty_acids(ukb_data)

}
\references{
Belkadi A. \emph{et al.} Size-resolved lipoprotein fatty acid content as a
novel nuclear magnetic resonance-derived trait specifically associates with
genetic variants that control fatty acid metabolism. \emph{J. Proteome Res.}
 (2026) doi: \href{https://pubs.acs.org/doi/10.1021/acs.jproteome.6c00107}{10.1021/acs.jproteome.6c00107}
}
\seealso{
\code{\link{nmr_info}} for list of computed lipoprotein fatty acid
  biomarkers, \code{\link{compute_lipoprotein_fatty_acids_flags}()} for
  obtaining an aggregate of the biomarker QC flags from the biomarkers
  underlying each computed lipoprotein fatty acid, and
  \code{\link{extract_biomarkers}()} for details on how raw field data
  extracted by the Table Exporter tool is processed.
}
