% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qc_flags_user.R
\name{compute_lipoprotein_fatty_acids_flags}
\alias{compute_lipoprotein_fatty_acids_flags}
\title{Aggregate QC Flags for the fatty acids computed in each lipoprotein class and subclass}
\usage{
compute_lipoprotein_fatty_acids_flags(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.
May either be raw field data output by the Table Exporter tool on the UK
Biobank Research Analysis Platform or data with column names corresponding
to biomarkers listed in \code{\link{nmr_info}}.}
}
\value{
a \code{data.frame} with QC flags aggregated for the computed
 lipoprotein fatty acids biomarkers.
}
\description{
For the 17 biomarkers computed by \code{\link{compute_lipoprotein_fatty_acids}()},
aggregates the biomarker QC flags from the underlying esterified cholesterol,
phospholipids, and triglyceride components, which can be useful for
determining the reason underlying missing values in the computed lipoprotein
fatty acid biomarkers.
}
\details{
If your UK Biobank project only has access to a subset of biomarkers, then
this function will only return the subset of ratios that can be computed from
the biomarker data provided.

Biomarker QC Flags in the input data are also returned alongside those
aggregated by this function for the computed biomarker ratios.
}
\examples{
ukb_data <- ukbnmr::test_data # Toy example dataset for testing package
biomarker_qc_flags <- compute_lipoprotein_fatty_acids_flags(ukb_data)

}
\seealso{
\code{\link{nmr_info}} for list of computed lipoprotein fatty acid
  biomarkers, \code{\link{compute_lipoprotein_fatty_acids_flags}()} for
  obtaining an aggregate of the biomarker QC flags from the biomarkers
  underlying each computed lipoprotein fatty acid, and
  \code{\link{extract_biomarkers}()} for details on how raw field data
  extracted by the Table Exporter tool is processed.
}
