% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{compute_nightingale_ratios}
\alias{compute_nightingale_ratios}
\title{Compute 81 biomarker ratios on the Nightingale platform}
\usage{
compute_nightingale_ratios(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.}
}
\value{
a \code{data.frame} or \code{data.table} with column names "eid" and
       "visit_index" followed by columns for each of the 249 biomarkers and
       ratios.
}
\description{
The Nightingale Health NMR metabolomics biomarker platform quantifies
\href{https://nightingalehealth.com/biomarkers}{249 biomarkers}, including 81
biomarker ratios. UK Biobank currently only provides the 168 biomarkers that
are not ratios for \href{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=220}{download}.
This function will compute the 81 missing ratios from the 168 downloadable
biomarkers in UKB.
}
\details{
If your UK Biobank project only has access to a subset of biomarkers, then
this function will only return the subset of ratios that can be computed from
the biomarker data provided.

All biomarkers in the input data are also returned alongside the ratios computed
by this function.
}
