% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_biomarkers.R
\name{extract_biomarkers}
\alias{extract_biomarkers}
\title{Extract Nightingale biomarkers from a data.frame of UK Biobank fields}
\usage{
extract_biomarkers(x)
}
\arguments{
\item{x}{\code{data.frame} with column names "eid" followed by extracted
fields e.g. "23474-0.0", "23474-1.0", \dots, "23467-1.0".}
}
\value{
a \code{data.frame} with column names "eid", "instance", and
 "array_index", followed by columns for each biomarker.
}
\description{
Given an input \code{data.frame} loaded from an dataset extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
extracts the \href{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=220}{UK Biobank fields}
corresponding to the
\href{https://nightingalehealth.com/biomarkers}{Nightingale Health NMR metabolomics biomarkers}
giving them short variable names as listed in the \code{\link{nmr_info}} information data sheet
available in this package.
}
\details{
Data sets extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
have one row per UKB biobank participant whose project specific sample
identifier is given in the first column named "eid". Columns following this
have the format "<field_id>-<instance>.<array_index>", where here <field_id>
corresponds to a biomarker of interest, e.g. 23474 for 3-Hydroxybutyrate,
<instance> corresponds to the assessment time point, e.g. 0 for baseline
assessment, 1 for first repeat visit, and <array_index> gives a number for
repeated measurements at the same time point.

In the returned \code{data.frame} there is single column for each biomarker,
with additional columns for the instance and array index. Rows are uniquely
identifiable by the combination of entries in columns "eid", "instance",
and "array index".

This function will also work with data extracted by the \code{ukbtools} R
package.
}
\seealso{
\code{\link{nmr_info}} for details on returned biomarkers.
}
