% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_qc_fields}
\alias{sample_qc_fields}
\title{Nightingale biomarker sample information}
\format{
A data table with 10 rows and 2 columns:
\describe{
  \item{Name}{Column name assigned to the sample information field}
  \item{UKB.Field.ID}{Field ID in UK Biobank, see \url{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=222}}
}
}
\usage{
sample_qc_fields
}
\description{
Contains details on the Nightingale biomarker sample QC information available
in UKB.
}
\keyword{datasets}
