% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios_qc_flags.R
\name{compute_nightingale_ratio_qc_flags}
\alias{compute_nightingale_ratio_qc_flags}
\title{Compute QC Flags for 81 biomarker ratios on the Nightingale platform}
\usage{
compute_nightingale_ratio_qc_flags(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.}
}
\value{
a \code{data.frame} or \code{data.table} with column names "eid" and
       "visit_index" followed by columns for each of the 249 biomarkers and
       ratios containing their respective QC Flags.
}
\description{
For the 81 biomarker ratios computed by \code{\link{compute_nightingale_ratios}()},
aggregates the biomarker QC flags from the biomarkers composing each ratio
(see \code{\link{nmr_info}}), which can be useful for determining the reason
underlying missing values in the biomarker ratios.
}
\details{
If your UK Biobank project only has access to a subset of biomarkers, then
this function will only return the subset of ratios that can be computed from
the biomarker data provided.

Biomarker QC Flags in the input data are also returned alongside those
aggregated by this function for the computed biomarker ratios.
}
