% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios_qc_flags.R
\name{compute_nightingale_ratio_qc_flags}
\alias{compute_nightingale_ratio_qc_flags}
\title{Compute QC Flags for 81 biomarker ratios on the Nightingale platform}
\usage{
compute_nightingale_ratio_qc_flags(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.}
}
\value{
a \code{data.frame} or \code{data.table} with column names "eid",
       "visit_index", and "repeat_index", followed by columns for each of the
        249 biomarkers and ratios.
}
\description{
For the 81 biomarker ratios computed by \code{\link{compute_nightingale_ratios}()},
aggregates the biomarker QC flags from the biomarkers composing each ratio
(see \code{\link{nmr_info}}), which can be useful for determining the reason
underlying missing values in the biomarker ratios.
}
\details{
Data sets extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
have one row per UKB biobank participant whose project specific sample
identifier is given in the first column named "eid". Columns following this
have the format "<field_id>-<visit_index>.<repeat_index>", where here <field_id>
corresponds to a biomarker of interest, e.g. 23774 for the QC Flag for 3-Hydroxybutyrate,
<visit_index> corresponds to the assessment time point, e.g. 0 for baseline
assessment, 1 for first repeat visit, and <repeat_index> gives a number for
repeated measurements at the same time point.

In the returned \code{data.frame} there is single column for each biomarker,
with additional columns for the visit and repeat index. Rows are uniquely
identifiable by the combination of entries in columns "eid" and "visit_index"
and "repeat_index".

Input data may be (1) a raw dataset extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
and loaded into R, "(2) a raw dataset extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
and loaded into R using the \code{ukbtools} R package, "or (3) a processed
\code{data.frame} of biomarker data obtained from the above using \code{\link{extract_biomarkers}()}.

A \code{data.table} will be returned instead of a \code{data.frame} if the
the user has loaded the package into their R session.
}
