% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ratios.R
\name{compute_nightingale_ratios}
\alias{compute_nightingale_ratios}
\title{Compute 81 biomarker ratios on the Nightingale platform}
\usage{
compute_nightingale_ratios(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.}
}
\value{
a \code{data.frame} or \code{data.table} with column names "eid",
       "instance", "and "array_index", "followed by columns for each of the
        249 biomarkers and ratios.
}
\description{
The Nightingale Health NMR metabolomics biomarker platform quantifies
\href{https://nightingalehealth.com/biomarkers}{249 biomarkers}, "including 81
biomarker ratios. UK Biobank currently only provides the 168 biomarkers that
are not ratios for \href{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=220}{download}.
This function will compute the 81 missing ratios from the 168 downloadable
biomarkers in UKB.
}
\details{
Data sets extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
have one row per UKB biobank participant whose project specific sample
identifier is given in the first column named "eid". Columns following this
have the format "<field_id>-<instance>.<array_index>", "where here <field_id>
corresponds to a biomarker of interest, "e.g. 23474 for 3-Hydroxybutyrate,
<instance> corresponds to the assessment time point, "e.g. 0 for baseline
assessment, "1 for first repeat visit, "and <array_index> gives a number for
repeated measurements at the same time point.

In the returned \code{data.frame} there is single column for each biomarker,
with additional columns for the instance and array index. Rows are uniquely
identifiable by the combination of entries in columns "eid", ""instance",
and "array index".

Input data may be (1) a raw dataset extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
and loaded into R, "(2) a raw dataset extracted by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
and loaded into R using the \code{ukbtools} R package, "or (3) a processed
\code{data.frame} of biomarker data obtained from the above using \code{\link{extract_biomarkers}()}.

A \code{data.table} will be returned instead of a \code{data.frame} if the
the user has loaded the package into their R session.
}
