% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_biomarkers_user.R
\name{compute_nightingale_ratios}
\alias{compute_nightingale_ratios}
\title{Compute 81 biomarker ratios on the Nightingale platform}
\usage{
compute_nightingale_ratios(x)
}
\arguments{
\item{x}{\code{data.frame} containing NMR metabolomics data from UK Biobank.
May either be raw field data output by
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
or data with column names corresponding to biomarkers listed in \code{\link{nmr_info}}.}
}
\value{
a \code{data.frame} with the additional computed biomarker ratios.
}
\description{
The Nightingale Health NMR metabolomics biomarker platform quantifies
\href{https://research.nightingalehealth.com/biomarkers/}{249 biomarkers}, including 81
biomarker ratios. Prior to the August 2021 update, UK Biobank only provided the 168
biomarkers that are not ratios for
\href{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=220}{download}.
This function will compute the 81 missing ratios.
}
\details{
If your UK Biobank project only has access to a subset of biomarkers, then
this function will only return the subset of ratios that can be computed from
the biomarker data provided.

All biomarkers in the input data are also returned alongside the ratios computed
by this function.
}
\examples{
\dontrun{
# Workflow for computing the 81 biomarker ratios missing prior to the August
# 2021 update of the UK Biobank NMR metabolomics biomarker data.
ukb_data <- fread("path/to/extracted.csv")
nmr <- compute_nightingale_ratios(ukb_data)
fwrite(nmr, file="path/to/nmr_biomarker_data.csv")

# You may also want to aggregate and save the quality control flags for each
# sample from the biomarkers underlying each ratio, adding them as additional
# columns to the input data (see help("compute_nightingale_ratio_qc_flags")).
biomarker_qc_flags <- compute_nightingale_ratio_qc_flags(nmr)
fwrite(biomarker_qc_flags, file="path/to/nmr_biomarker_qc_flags.csv")
}

}
\seealso{
\code{\link{nmr_info}} for list of computed biomarker ratios,
  \code{\link{compute_nightingale_ratio_qc_flags}()} for obtaining an
  aggregate of the biomarker QC flags from the biomarkers underlying each
  computed ratio, and \code{\link{extract_biomarkers}()} for details on how
  raw data from
  \href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{ukbconv}
  is processed.
}
